+++ Evaluator.php
@@ -26,28 +26,49 @@ class Evaluator implements ExpressionLanguageAwareInterface
 
     protected $context;
 
-    public function __construct(
-        $expression,
-        array $context = [],
-        bool $catchExceptions = false,
-        bool $errorToException = false
-    )
+    /**
+     * Invoke arguments context variables names. Default: arg0, arg1, etc.
+     */
+    protected $argNames;
+
+    public function __construct($expression, array $context = [], array $argNames = [])
     {
-        $this->context = $context;
-        $this->catchExceptions = $catchExceptions;
-        $this->errorToException = $errorToException;
         is_array($expression)
             ? $this->setExpressions($expression)
             : $this->addExpression($expression);
+
+        $this->context = $context;
+
+        foreach ($argNames as $name) {
+            if (!is_string($name)) {
+                throw new \UnexpectedValueException('String expected as argument name!');
+            }
+        }
+        $this->argNames = $argNames;
     }
 
     public function __invoke()
     {
         foreach (func_get_args() as $id => $arg) {
-            $this->context['arg'.$id] = $arg;
+            $this->context[$this->argNames[$id] ?? 'arg'.$id] = $arg;
         }
         $results = $this->evaluateExpressions($this->context);
 
         return end($results);
+    }
+
+    public function __call(string $name, array $arguments)
+    {
+        return $this(...$arguments);
+    }
+
+    public function catchExceptions()
+    {
+        $this->catchExceptions = true;
+    }
+
+    public function errorToException()
+    {
+        $this->errorToException = true;
     }
 }