+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # CHANGELOG
 
+## [1.2.0] - 2022-05-21
+
+### Changed
+
+- catch exceptions and error to exception options moved from constructor
+  to method
+
+### Added
+
+- optional mapping names of invoke arguments context variables inplace
+  of default `arg0`, `arg1`, etc.
+- catch any method call as evaluator invoke method
+
 ## [1.1.0] - 2022-05-16
 
 ### Added